chore: Refactoring sorting into separate files.
