+++ src/components/Table/GridTable.stories.tsx
@@ -238,257 +238,6 @@ export function NestedCards() {
   );
 }
 
-// Every row by-definition opens or closes a card.
-// Every row has a space between.
-
-export function NestedCardsProofOfConcept() {
-  // Combine rows into a single "kind: chrome" div that goes between
-  // each real row.
-
-  return (
-    <div css={Css.bgGray900.p2.$}>
-      <div
-        css={{
-          ...Css.dg.gtc("100px 100px 100px").$,
-          // All open/close rows are 4px
-          "& > [data-card-open]": Css.add({ gridColumn: "span 3" }).$,
-          "& > [data-card-close]": Css.add({ gridColumn: "span 3" }).$,
-          "& > [data-card-open] div": Css.hPx(4).px1.$,
-          "& > [data-card-close] div": Css.hPx(4).px1.$,
-          // open/close corners
-          "& > div[data-card-open='level1'] > div": Css.brt4.$,
-          "& > div[data-card-open='level2'] > div > div": Css.brt4.bt.br.bl.bGray200.$,
-          "& > div[data-card-open='level3'] > div > div > div": Css.brt4.$,
-          "& > div[data-card-close='level3'] > div > div > div": Css.brb4.$,
-          "& > div[data-card-close='level2'] > div > div": Css.brb4.bb.br.bl.bGray200.$,
-          "& > div[data-card-close='level1'] > div": Css.brb4.$,
-          // the 2nd level is special/has a border
-          "& > div[data-card-open='level3'] > div > div": Css.bGray200.bl.br.$,
-          "& > div[data-card-close='level3'] > div > div": Css.bGray200.bl.br.$,
-          // borders (TODO should not apply to 1st level)
-          "& > div[data-card] > div:first-of-type > div": Css.bl.bGray200.$,
-          "& > div[data-card] > div:last-of-type > div": Css.br.bGray200.$,
-          // spacers
-          "& > div[data-spacer]": Css.add({ gridColumn: "span 3" }).hPx(8).$,
-          "& > div[data-spacer] div": Css.hPx(8).px1.$,
-          // the 2nd level is special/has a border
-          "& > div[data-spacer] > div > div": Css.bl.br.bGray200.$,
-          // backgrounds
-          "& div[data-level='level1']": Css.bgGray100.$,
-          "& div[data-level='level2']": Css.bgWhite.$,
-          "& div[data-level='level3']": Css.bgGreen200.$,
-        }}
-      >
-        {/* Grand-parent row */}
-        <div data-card-open="level1">
-          <div data-level="level1" />
-        </div>
-        <div data-card="level1" css={Css.display("contents").$}>
-          <div data-level="level1" css={Css.pl1.$}>
-            Milestone 1
-          </div>
-          <div data-level="level1">Milestone 1</div>
-          <div data-level="level1" css={Css.pr1.$}>
-            Milestone 1
-          </div>
-        </div>
-
-        {/*Row is not closing, so spacer level1.*/}
-        <div data-spacer="true">
-          <div data-level="level1" />
-        </div>
-
-        {/* Child row */}
-        <div data-card-open="level2">
-          <div data-level="level1">
-            <div data-level="level2" />
-          </div>
-        </div>
-        <div data-card="level2" css={Css.display("contents").$}>
-          <div data-level="level1" css={Css.pl1.$}>
-            <div data-level="level2" css={Css.pl1.$}>
-              Group 1
-            </div>
-          </div>
-          <div data-level="level2">Group 1</div>
-          <div data-level="level1" css={Css.pr1.$}>
-            <div data-level="level2" css={Css.pr1.$}>
-              Group 1
-            </div>
-          </div>
-        </div>
-        <div data-card-close="level2">
-          <div data-level="level1">
-            <div data-level="level2" />
-          </div>
-        </div>
-
-        {/*Row has closed, so spacer level1.*/}
-        <div data-spacer="true">
-          <div data-level="level1" />
-        </div>
-
-        {/* Child row */}
-        <div data-card-open="level2">
-          <div data-level="level1">
-            <div data-level="level2" />
-          </div>
-        </div>
-        <div data-card="level2" css={Css.display("contents").$}>
-          <div data-level="level1" css={Css.pl1.$}>
-            <div data-level="level2" css={Css.pl1.$}>
-              Group 2
-            </div>
-          </div>
-          <div data-level="level2">Group 2</div>
-          <div data-level="level1" css={Css.pr1.$}>
-            <div data-level="level2" css={Css.pr1.$}>
-              Group 2
-            </div>
-          </div>
-        </div>
-
-        {/*Row is not closing, so spacer level2.*/}
-        <div data-spacer="true">
-          <div data-level="level1">
-            <div data-level="level2" />
-          </div>
-        </div>
-
-        {/* 1st Grandchild row. */}
-        <div data-card-open="level3">
-          <div data-level="level1">
-            <div data-level="level2">
-              <div data-level="level3" />
-            </div>
-          </div>
-        </div>
-        <div data-card="level3" css={Css.display("contents").$}>
-          <div data-level="level1" css={Css.pl1.$}>
-            <div data-level="level2" css={Css.pl1.$}>
-              <div data-level="level3" css={Css.pl1.$}>
-                Task 1
-              </div>
-            </div>
-          </div>
-          <div data-level="level3">Task 1</div>
-          <div data-level="level1" css={Css.pr1.$}>
-            <div data-level="level2" css={Css.pr1.$}>
-              <div data-level="level3" css={Css.pr1.$}>
-                Task 1
-              </div>
-            </div>
-          </div>
-        </div>
-        <div data-card-close="level3">
-          <div data-level="level1">
-            <div data-level="level2">
-              <div data-level="level3" />
-            </div>
-          </div>
-        </div>
-
-        {/*Row is not closing, so spacer level2.*/}
-        <div data-spacer="true">
-          <div data-level="level1">
-            <div data-level="level2" />
-          </div>
-        </div>
-
-        {/* 2nd Grandchild row. */}
-        <div data-card-open="level3">
-          <div data-level="level1">
-            <div data-level="level2">
-              <div data-level="level3" />
-            </div>
-          </div>
-        </div>
-        <div data-card="level3" css={Css.display("contents").$}>
-          <div data-level="level1" css={Css.pl1.$}>
-            <div data-level="level2" css={Css.pl1.$}>
-              <div data-level="level3" css={Css.pl1.$}>
-                Task 2
-              </div>
-            </div>
-          </div>
-          <div data-level="level3">Task 2</div>
-          <div data-level="level1" css={Css.pr1.$}>
-            <div data-level="level2" css={Css.pr1.$}>
-              <div data-level="level3" css={Css.pr1.$}>
-                Task 2
-              </div>
-            </div>
-          </div>
-        </div>
-        <div data-card-close="level3">
-          <div data-level="level1">
-            <div data-level="level2">
-              <div data-level="level3" />
-            </div>
-          </div>
-        </div>
-        <div data-card-close="level2">
-          <div data-level="level1">
-            <div data-level="level2" />
-          </div>
-        </div>
-
-        {/*Child has closed, so spacer level1.*/}
-        <div data-spacer="true">
-          <div data-level="level1" />
-        </div>
-
-        {/* Child row, could be "last grand-child" as well as "last child" */}
-        <div data-card-open="level2">
-          <div data-level="level1">
-            <div data-level="level2" />
-          </div>
-        </div>
-        <div data-card="level2" css={Css.display("contents").$}>
-          <div data-level="level1" css={Css.pl1.$}>
-            <div data-level="level2" css={Css.pl1.$}>
-              Group 3
-            </div>
-          </div>
-          <div data-level="level2">Group 3</div>
-          <div data-level="level1" css={Css.pr1.$}>
-            <div data-level="level2" css={Css.pr1.$}>
-              Group 3
-            </div>
-          </div>
-        </div>
-        <div data-card-close="level2">
-          <div data-level="level1">
-            <div data-level="level2" />
-          </div>
-        </div>
-        <div data-card-close="level1">
-          <div data-level="level1" />
-        </div>
-
-        {/*Child has closed, so spacer level0.*/}
-        <div data-spacer="true" />
-
-        <div data-card-open="level1">
-          <div data-level="level1" />
-        </div>
-        <div data-card="level1" css={Css.display("contents").$}>
-          <div data-level="level1" css={Css.pl1.$}>
-            Milestone 2
-          </div>
-          <div data-level="level1">Milestone 2</div>
-          <div data-level="level1" css={Css.pr1.$}>
-            Milestone 2
-          </div>
-        </div>
-        <div data-card-close="level1">
-          <div data-level="level1" />
-        </div>
-      </div>
-    </div>
-  );
-}
-
 export function OneOffInlineTable() {
   const items: { code: string; name: string; quantity: number }[] = [
     { code: "AAA", name: "Aaa", quantity: 1 },
